+++ src/rx-query.ts
@@ -221,18 +221,24 @@ export class RxQueryBase<
      * executes the query on the database
      * @return results-array with document-data
      */
-    async _execOverDatabase(): Promise<RxDocumentData<RxDocType>[] | number> {
+    async _execOverDatabase(): Promise<{
+        result: RxDocumentData<RxDocType>[] | number;
+        counter: number;
+    }> {
         this._execOverDatabaseCount = this._execOverDatabaseCount + 1;
         if (this.op === 'count') {
             const preparedQuery = this.getPreparedQuery();
-            const result = await this.collection.storageInstance.count(preparedQuery);
-            if (result.mode === 'slow' && !this.collection.database.allowSlowCount) {
+            const countResult = await this.collection.storageInstance.count(preparedQuery);
+            if (countResult.mode === 'slow' && !this.collection.database.allowSlowCount) {
                 throw newRxError('QU14', {
                     collection: this.collection,
                     queryObj: this.mangoQuery
                 });
             } else {
-                return result.count;
+                return {
+                    result: countResult.count,
+                    counter: this.collection._changeEventBuffer.getCounter()
+                };
             }
         }
 
@@ -260,14 +266,17 @@ export class RxQueryBase<
                     ret.set(doc.primary, doc);
                 });
             }
-            return ret as any;
+            return {
+                result: ret as any,
+                counter: this.collection._changeEventBuffer.getCounter()
+            };
         }
 
-
-        const docsPromise = queryCollection<RxDocType>(this as any);
-        return docsPromise.then(docs => {
-            return docs;
-        });
+        const result = await queryCollection<RxDocType>(this as any);
+        return {
+            result: result.docs,
+            counter: result.counter
+        };
     }
 
     /**
@@ -523,14 +532,6 @@ async function _ensureEqual(rxQuery: RxQueryBase<any, any>): Promise<boolean> {
         await Promise.all(Array.from(rxQuery.collection.awaitBeforeReads).map(fn => fn()));
     }
 
-    // Optimisation shortcut
-    if (
-        rxQuery.collection.database.closed ||
-        _isResultsInSync(rxQuery)
-    ) {
-        return false;
-    }
-
     rxQuery._ensureEqualQueue = rxQuery._ensureEqualQueue
         .then(() => __ensureEqual(rxQuery));
     return rxQuery._ensureEqualQueue;
@@ -572,7 +573,6 @@ function __ensureEqual<RxDocType>(rxQuery: RxQueryBase<RxDocType, any>): Promise
             mustReExec = true;
         } else {
             rxQuery._latestChangeEvent = rxQuery.asRxQuery.collection._changeEventBuffer.getCounter();
-
             const runChangeEvents: RxStorageChangeEvent<RxDocType>[] = rxQuery.asRxQuery.collection
                 ._changeEventBuffer
                 .reduceByLastOfDoc(missedChangeEvents);
@@ -617,14 +617,15 @@ function __ensureEqual<RxDocType>(rxQuery: RxQueryBase<RxDocType, any>): Promise
     // oh no we have to re-execute the whole query over the database
     if (mustReExec) {
         return rxQuery._execOverDatabase()
-            .then(newResultData => {
+            .then(result => {
+                const newResultData = result.result;
 
                 /**
                  * The RxStorage is defined to always first emit events and then return
                  * on bulkWrite() calls. So here we have to use the counter AFTER the execOverDatabase()
                  * has been run, not the one from before.
                  */
-                rxQuery._latestChangeEvent = rxQuery.collection._changeEventBuffer.getCounter();
+                rxQuery._latestChangeEvent = result.counter;
 
                 // A count query needs a different has-changed check.
                 if (typeof newResultData === 'number') {
@@ -663,7 +664,15 @@ function __ensureEqual<RxDocType>(rxQuery: RxQueryBase<RxDocType, any>): Promise
  */
 export async function queryCollection<RxDocType>(
     rxQuery: RxQuery<RxDocType> | RxQueryBase<RxDocType, any>
-): Promise<RxDocumentData<RxDocType>[]> {
+): Promise<{
+    docs: RxDocumentData<RxDocType>[];
+    /**
+     * We need to remember the counter directly here
+     * because getting if after the returned Promise is resolved,
+     * can result in a value that no longer matches the result set.
+     */
+    counter: number;
+}> {
     let docs: RxDocumentData<RxDocType>[] = [];
     const collection = rxQuery.collection;
 
@@ -714,7 +723,10 @@ export async function queryCollection<RxDocType>(
         const queryResult = await collection.storageInstance.query(preparedQuery);
         docs = queryResult.documents;
     }
-    return docs;
+    return {
+        docs,
+        counter: collection._changeEventBuffer.getCounter()
+    };
 
 }